+++ src/main/java/com/exercise/basic/SoHoc/Bai1.java
@@ -1,4 +1,4 @@
-package com.anhtuan.exercise.basic;
+package com.exercise.basic.SoHoc;
 
 public class Bai1 {
